+++ buildSrc/src/main/java/Config.kt
@@ -32,6 +32,10 @@ object Config {
         val compileSdkVersion = sdkVersion
 
         val abiFilters = listOf("x86", "armeabi-v7a", "x86_64", "arm64-v8a")
+
+        fun shouldSkipDebugVariant(name: String): Boolean {
+            return System.getenv("CI")?.toBoolean() ?: false && name == "debug"
+        }
     }
 
     object Libs {
@@ -69,7 +73,7 @@ object Config {
         val springSecurityConfig = "org.springframework.security:spring-security-config"
         val springAop = "org.springframework:spring-aop"
         val aspectj = "org.aspectj:aspectjweaver"
-        val servletApi = "javax.servlet:javax.servlet-api"
+        val servletApi = "javax.servlet:javax.servlet-api:3.1.0"
 
         val apacheHttpClient = "org.apache.httpcomponents.client5:httpclient5:5.0.4"
 
@@ -83,6 +87,10 @@ object Config {
         val fragment = "androidx.fragment:fragment-ktx:1.3.5"
 
         val reactorCore = "io.projectreactor:reactor-core:3.4.6"
+
+        private val feignVersion = "11.6"
+        val feignCore = "io.github.openfeign:feign-core:$feignVersion"
+        val feignGson = "io.github.openfeign:feign-gson:$feignVersion"
     }
 
     object AnnotationProcessors {
@@ -122,8 +130,9 @@ object Config {
         // use RC2 to drop jcenter because of kotlinx-html
         val detektVersion = "1.17.1"
         val detektPlugin = "io.gitlab.arturbosch.detekt"
-        val binaryCompatibilityValidatorPlugin = "org.jetbrains.kotlinx:binary-compatibility-validator:0.5.0"
-        val binaryCompatibilityValidator = "binary-compatibility-validator"
+        val binaryCompatibilityValidatorVersion = "0.6.0"
+        val binaryCompatibilityValidatorPlugin = "org.jetbrains.kotlinx:binary-compatibility-validator:$binaryCompatibilityValidatorVersion"
+        val binaryCompatibilityValidator = "org.jetbrains.kotlinx.binary-compatibility-validator"
     }
 
     object Sentry {